+++ CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.3.2](https://github.com/@nativescript-community/ui-collectionview/compare/v5.3.1...v5.3.2) (2023-10-20)
+
+### Bug Fixes
+
+* **android:** disable update animation ([ed97594](https://github.com/@nativescript-community/ui-collectionview/commit/ed97594ed6238bcb0da26a9102b024780d454cde))
+* **android:** ignore recycled holders in `eachChild` calls ([996bee4](https://github.com/@nativescript-community/ui-collectionview/commit/996bee4fd73510ad17db239aeb6cd02b8725f34d))
+
+## [5.3.1](https://github.com/@nativescript-community/ui-collectionview/compare/v5.3.0...v5.3.1) (2023-10-18)
+
+**Note:** Version bump only for package ui-collectionview
+
+# [5.3.0](https://github.com/@nativescript-community/ui-collectionview/compare/v5.2.0...v5.3.0) (2023-10-18)
+
+### Bug Fixes
+
+* a bunch of fixes to make collectionview faster ([cc550d7](https://github.com/@nativescript-community/ui-collectionview/commit/cc550d74626dfe58ff6c593cfc89b3c189771ded))
+* **android:** regression fixes ([fda4f74](https://github.com/@nativescript-community/ui-collectionview/commit/fda4f74b47ac7a0e21d351f3c600828c1c69a2dd))
+* **android:** removed animation lib for now ([b0343f8](https://github.com/@nativescript-community/ui-collectionview/commit/b0343f85fdd2ca5d415bdc7fffd3467ecff0f12c))
+* vue3 component improvement. Should be much faster ([7bc0498](https://github.com/@nativescript-community/ui-collectionview/commit/7bc0498586db2e931635b62706b26d8eca608284))
+
+### Features
+
+* **android:** layoutCompleted event ([3a239a3](https://github.com/@nativescript-community/ui-collectionview/commit/3a239a3473de25f8d287ea9e7f71d3a2a7449dac))
+
 # [5.2.0](https://github.com/@nativescript-community/ui-collectionview/compare/v5.1.9...v5.2.0) (2023-09-12)
 
 ### Bug Fixes