+++ packages/swipemenu/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.3.2](https://github.com/nativescript-community/ui-collectionview/compare/v5.3.1...v5.3.2) (2023-10-20)
+
+**Note:** Version bump only for package @nativescript-community/ui-collectionview-swipemenu
+
+## [5.3.1](https://github.com/nativescript-community/ui-collectionview/compare/v5.3.0...v5.3.1) (2023-10-18)
+
+**Note:** Version bump only for package @nativescript-community/ui-collectionview-swipemenu
+
+# [5.3.0](https://github.com/nativescript-community/ui-collectionview/compare/v5.2.0...v5.3.0) (2023-10-18)
+
+**Note:** Version bump only for package @nativescript-community/ui-collectionview-swipemenu
+
 # [5.2.0](https://github.com/nativescript-community/ui-collectionview/compare/v5.1.9...v5.2.0) (2023-09-12)
 
 **Note:** Version bump only for package @nativescript-community/ui-collectionview-swipemenu